Extract Codex hook runtime
