+++ src/adapters/codex-hook-runtime.ts
@@ -1,0 +1,297 @@
+import { somaPolicyPrivateMarkers } from "../policy";
+
+function renderCodexPolicyHookRuntime(somaHome: string, homeDir?: string): string[] {
+  const policyMarkers = somaPolicyPrivateMarkers(somaHome, homeDir);
+
+  return [
+    `const SOMA_POLICY_MARKERS = ${JSON.stringify(policyMarkers)};`,
+    "",
+    "function hasSomaPolicyMarker(content) {",
+    '  return SOMA_POLICY_MARKERS.some((marker) => (content || "").includes(marker));',
+    "}",
+    "",
+    "function policyRelevantContent(content) {",
+    '  return (content || "").split("\\n").filter((line) => hasSomaPolicyMarker(line)).join("\\n");',
+    "}",
+    "",
+    "function runSomaPolicyCheck(targets) {",
+    '  const args = ["run", "soma", "policy", "check", "--soma-home", SOMA_HOME, "--substrate", "codex", "--action", "write", "--targets-env", "SOMA_POLICY_TARGETS", "--record", "deny", "--json"];',
+    "",
+    '  return spawnSync("bun", args, {',
+    "    cwd: somaRepoPath(),",
+    '    encoding: "utf8",',
+    "    timeout: 25000,",
+    "    env: { ...process.env, SOMA_POLICY_TARGETS: JSON.stringify(targets) }",
+    "  });",
+    "}",
+    "",
+    "function denyPreToolUse(reason) {",
+    "  console.log(JSON.stringify({",
+    "    hookSpecificOutput: {",
+    '      hookEventName: "PreToolUse",',
+    '      permissionDecision: "deny",',
+    "      permissionDecisionReason: reason",
+    "    }",
+    "  }));",
+    "  process.exit(0);",
+    "}",
+  ];
+}
+
+export function renderCodexLifecycleHook(somaHome: string, homeDir?: string): string {
+  return [
+    "#!/usr/bin/env node",
+    'import { spawnSync } from "node:child_process";',
+    'import { readFileSync, writeFileSync } from "node:fs";',
+    'import { isAbsolute, resolve } from "node:path";',
+    "",
+    `const SOMA_HOME = ${JSON.stringify(somaHome)};`,
+    "",
+    "function readHookInput() {",
+    "  try {",
+    '    return JSON.parse(readFileSync(0, "utf8"));',
+    "  } catch {",
+    "    return {};",
+    "  }",
+    "}",
+    "",
+    "function somaRepoPath() {",
+    "  if (process.env.SOMA_REPO) return process.env.SOMA_REPO;",
+    "  try {",
+    '    return readFileSync(`${process.env.HOME}/.codex/memories/soma/soma-repo.txt`, "utf8").trim();',
+    "  } catch {",
+    "    return process.cwd();",
+    "  }",
+    "}",
+    "",
+    "function runSomaLifecycle(event, sessionId) {",
+    '  const args = ["run", "soma", "lifecycle", event, "--soma-home", SOMA_HOME, "--substrate", "codex"];',
+    "  if (sessionId) {",
+    '    args.push("--session-id", sessionId);',
+    "  }",
+    "",
+    '  return spawnSync("bun", args, { cwd: somaRepoPath(), encoding: "utf8", timeout: 25000 });',
+    "}",
+    "",
+    "function runSomaClassification(prompt) {",
+    '  return spawnSync("bun", ["run", "soma", "algorithm", "classify", "--prompt", prompt || ""], {',
+    "    cwd: somaRepoPath(),",
+    '    encoding: "utf8",',
+    "    timeout: 25000",
+    "  });",
+    "}",
+    "",
+    ...renderCodexPolicyHookRuntime(somaHome, homeDir),
+    "",
+    "function resolveToolPath(path, cwd) {",
+    "  return isAbsolute(path) ? path : resolve(cwd || process.cwd(), path);",
+    "}",
+    "",
+    "function pushPatchTarget(targets, target) {",
+    "  if (target) targets.push({ filePath: target.filePath, content: target.lines.filter((line) => hasSomaPolicyMarker(line)).join(\"\\n\") });",
+    "}",
+    "",
+    "function extractPatchTargets(patch, cwd) {",
+    "  const targets = [];",
+    "  let current;",
+    '  const pattern = /^\\*\\*\\* (?:Add|Update|Delete) File: (.+)$/;',
+    '  const movePattern = /^\\*\\*\\* Move to: (.+)$/;',
+    '  for (const line of (patch || "").split("\\n")) {',
+    "    const fileMatch = line.match(pattern) || line.match(movePattern);",
+    "    if (fileMatch) {",
+    "      pushPatchTarget(targets, current);",
+    "      current = { filePath: resolveToolPath(fileMatch[1].trim(), cwd), lines: [] };",
+    "      continue;",
+    "    }",
+    '    if (current && line.startsWith("+") && !line.startsWith("+++")) {',
+    "      current.lines.push(line.slice(1));",
+    "    }",
+    "  }",
+    "  pushPatchTarget(targets, current);",
+    "  return targets;",
+    "}",
+    "",
+    "function extractWriteTargets(input) {",
+    "  const toolName = input.tool_name || input.toolName || \"\";",
+    "  const toolInput = input.tool_input || input.toolInput || {};",
+    "  const cwd = input.cwd || process.cwd();",
+    "  const filePath = typeof toolInput === \"object\" ? resolveToolPath(toolInput.file_path || toolInput.filePath || cwd, cwd) : cwd;",
+    "",
+    '  if (toolName === "Write") {',
+    "    return [{ filePath, content: toolInput.content || \"\" }];",
+    "  }",
+    "",
+    '  if (toolName === "Edit") {',
+    "    return [{ filePath, content: toolInput.new_string || toolInput.newString || \"\" }];",
+    "  }",
+    "",
+    '  if (toolName === "MultiEdit") {',
+    "    const edits = Array.isArray(toolInput.edits) ? toolInput.edits : [];",
+    "    return edits.map((edit) => ({ filePath, content: edit?.new_string || edit?.newString || \"\" }));",
+    "  }",
+    "",
+    '  if (toolName === "apply_patch") {',
+    "    const content = typeof toolInput === \"string\" ? toolInput : toolInput.patch || toolInput.command || JSON.stringify(toolInput);",
+    "    const targets = extractPatchTargets(content, cwd);",
+    "    return targets.length > 0 ? targets : [{ filePath: cwd, content: policyRelevantContent(content) }];",
+    "  }",
+    "",
+    "  return [];",
+    "}",
+    "",
+    "function parseClassification(output) {",
+    "  const fields = {};",
+    '  for (const line of output.split("\\n")) {',
+    '    const separator = line.indexOf(": ");',
+    "    if (separator === -1) continue;",
+    "    fields[line.slice(0, separator)] = line.slice(separator + 2).trim();",
+    "  }",
+    "  return fields;",
+    "}",
+    "",
+    "function renderPromptClassificationContext(classification) {",
+    '  const mode = (classification.mode || "algorithm").toUpperCase();',
+    '  const effort = classification.effort && classification.effort !== "none" ? classification.effort : "";',
+    '  const source = classification.source || "unknown";',
+    '  const reason = classification.reason || "No reason provided.";',
+    "  const lines = [",
+    '    "# Soma Prompt Classification",',
+    "    `MODE: ${mode}`,",
+    "    effort ? `TIER: ${effort}` : undefined,",
+    "    `SOURCE: ${source}`,",
+    "    `REASON: ${reason}`,",
+    '    ""',
+    "  ].filter(Boolean);",
+    "",
+    '  if (mode === "ALGORITHM") {',
+    "    lines.push(",
+    '      "Operating requirement:",',
+    '      "- Use the `the-algorithm` skill for this turn.",',
+    '      "- Before giving a substantive answer, create or update a Soma Algorithm run unless this prompt only continues an already-active run.",',
+    '      "- Use the classified TIER as the effort level unless the user explicitly overrides it.",',
+    '      "- Preserve verification evidence in the run before declaring the answer complete."',
+    "    );",
+    "  } else {",
+    "    lines.push(",
+    '      "Operating requirement:",',
+    '      "- This prompt may stay outside the full Algorithm harness unless conversation context makes it Algorithm-shaped."',
+    "    );",
+    "  }",
+    "",
+    '  return lines.join("\\n");',
+    "}",
+    "",
+    "function writeProjectedStartupContext(output) {",
+    '  const marker = "# Soma Startup Context";',
+    "  const index = output.indexOf(marker);",
+    "  if (index === -1) return undefined;",
+    "  const context = output.slice(index).trim();",
+    '  writeFileSync(`${process.env.HOME}/.codex/memories/soma/startup-context.md`, `${context}\\n`, "utf8");',
+    "  return context;",
+    "}",
+    "",
+    "function readProjectedStartupContext() {",
+    "  try {",
+    '    return readFileSync(`${process.env.HOME}/.codex/memories/soma/startup-context.md`, "utf8").trim();',
+    "  } catch {",
+    "    return undefined;",
+    "  }",
+    "}",
+    "",
+    "const event = process.argv[2];",
+    "const input = readHookInput();",
+    "",
+    'if (event === "pre-tool-use") {',
+    "  const targets = extractWriteTargets(input).filter((target) => hasSomaPolicyMarker(target.content));",
+    "  if (targets.length > 0) {",
+    "    const result = runSomaPolicyCheck(targets);",
+    "    const output = result.stdout || result.stderr || \"\";",
+    "    if (result.status !== 0) {",
+    "      denyPreToolUse(`Soma policy check failed closed: ${output || \"unknown error\"}`);",
+    "    }",
+    "    let policy;",
+    "    try {",
+    "      policy = JSON.parse(output);",
+    "    } catch {",
+    "      denyPreToolUse(`Soma policy check returned invalid JSON: ${output || \"empty output\"}`);",
+    "    }",
+    '    if (policy.decision === "deny") {',
+    "      const reason = policy.reason || \"Soma private context policy denied this write.\";",
+    "      denyPreToolUse(reason);",
+    "    }",
+    "  }",
+    "  console.log(JSON.stringify({ continue: true }));",
+    "  process.exit(0);",
+    "}",
+    "",
+    'if (event === "prompt-submit") {',
+    "  const result = runSomaClassification(input.prompt);",
+    "  if (result.status !== 0) {",
+    "    console.log(JSON.stringify({",
+    "      continue: true,",
+    "      hookSpecificOutput: {",
+    '        hookEventName: "UserPromptSubmit",',
+    '        additionalContext: `Soma prompt classification failed; if this prompt is substantial, use the-algorithm manually. ${result.stderr || result.stdout || ""}`',
+    "      }",
+    "    }));",
+    "    process.exit(0);",
+    "  }",
+    "  const context = renderPromptClassificationContext(parseClassification(result.stdout));",
+    "  console.log(JSON.stringify({",
+    "    continue: true,",
+    "    hookSpecificOutput: {",
+    '      hookEventName: "UserPromptSubmit",',
+    "      additionalContext: context",
+    "    }",
+    "  }));",
+    "  process.exit(0);",
+    "}",
+    "",
+    "const result = runSomaLifecycle(event, input.session_id);",
+    "",
+    "if (result.status !== 0) {",
+    '  if (event === "session-start") {',
+    "    const context = readProjectedStartupContext();",
+    "    console.log(JSON.stringify({",
+    "      continue: true,",
+    "      systemMessage: `Soma lifecycle hook fell back to projected context: ${result.stderr || result.stdout || \"unknown error\"}`,",
+    "      hookSpecificOutput: {",
+    '        hookEventName: "SessionStart",',
+    '        additionalContext: context || "Soma lifecycle context is unavailable; read ~/.codex/memories/soma/ when needed."',
+    "      }",
+    "    }));",
+    "    process.exit(0);",
+    "  }",
+    "  console.log(JSON.stringify({",
+    "    continue: true,",
+    '    systemMessage: `Soma lifecycle hook failed: ${result.stderr || result.stdout || "unknown error"}`',
+    "  }));",
+    "  process.exit(0);",
+    "}",
+    "",
+    'if (event === "session-start") {',
+    "  const context = writeProjectedStartupContext(result.stdout);",
+    "  console.log(JSON.stringify({",
+    "    continue: true,",
+    "    hookSpecificOutput: {",
+    '      hookEventName: "SessionStart",',
+    '      additionalContext: context || "Soma lifecycle context is available in ~/.codex/memories/soma/startup-context.md."',
+    "    }",
+    "  }));",
+    "  process.exit(0);",
+    "}",
+    "",
+    'if (event === "algorithm-updated") {',
+    "  console.log(JSON.stringify({",
+    "    continue: true,",
+    "    hookSpecificOutput: {",
+    '      hookEventName: "PostToolUse",',
+    '      additionalContext: "Soma Algorithm work index refreshed."',
+    "    }",
+    "  }));",
+    "  process.exit(0);",
+    "}",
+    "",
+    "console.log(JSON.stringify({ continue: true }));",
+  ].join("\n");
+}
